quick_bot-dev/requirements.txt
Flask==1.0.2
scikit-learn==0.20.3
langdetect==1.0.7
mecab-python3==0.7
numpy==1.15.0
requests==2.21.0